Migrate Parser type helpers to their own module.
