+++ npm-packages/meteor-vite/src/meteor/package/ParserTypes.ts
@@ -1,0 +1,34 @@
+import {FunctionExpression, ObjectExpression, ObjectProperty, StringLiteral} from "@babel/types";
+
+type KnownObjectProperty<TValue extends Pick<ObjectProperty, 'key' | 'value'>> = Omit<ObjectProperty, 'key' | 'value'> & TValue;
+type KnownObjectExpression<TValue extends Pick<ObjectExpression, 'properties'>> = Omit<ObjectExpression, 'properties'> & TValue;
+
+export type MeteorPackageProperty = KnownObjectProperty<{
+    key: StringLiteral, // File name
+    value: FunctionExpression | MeteorNestedPackageProperty, // Module function
+}>
+type MeteorNestedPackageProperty = KnownObjectExpression<{
+    properties: MeteorPackageProperty[]
+}>
+
+export const KnownModuleMethods = ['export', 'link', 'exportDefault'] as const;
+export type ModuleMethod = typeof KnownModuleMethods[number];
+
+export type PackageConfig = KnownObjectExpression<{
+    properties: [KnownObjectProperty<{
+        key: StringLiteral & { value: 'node_modules' }
+        value: KnownObjectExpression<{
+            properties: [KnownObjectProperty<{
+                key: StringLiteral & { value: 'meteor' },
+                value: KnownObjectExpression<{
+                    properties: [KnownObjectProperty<{
+                        key: StringLiteral, // Package name
+                        value: KnownObjectExpression<{
+                            properties: MeteorPackageProperty[]
+                        }>
+                    }>]
+                }>
+            }>]
+        }>
+    }>]
+}>
